+++ crates/miden-standards/src/account/faucets/fungible/mod.rs
@@ -596,8 +596,10 @@ pub fn create_multisig_user_fungible_faucet(
     token_policy_manager: TokenPolicyManager,
     account_type: AccountType,
 ) -> Result<Account, FungibleFaucetError> {
+    let asset_callbacks = AssetCallbackFlag::from(token_policy_manager.has_transfer_policy());
     AccountBuilder::new(init_seed)
         .account_type(account_type)
+        .with_asset_callbacks(asset_callbacks)
         .with_component(auth_component)
         .with_component(faucet)
         .with_component(Authority::AuthControlled)
@@ -616,8 +618,10 @@ pub fn create_guarded_user_fungible_faucet(
     token_policy_manager: TokenPolicyManager,
     account_type: AccountType,
 ) -> Result<Account, FungibleFaucetError> {
+    let asset_callbacks = AssetCallbackFlag::from(token_policy_manager.has_transfer_policy());
     AccountBuilder::new(init_seed)
         .account_type(account_type)
+        .with_asset_callbacks(asset_callbacks)
         .with_component(auth_component)
         .with_component(faucet)
         .with_component(Authority::AuthControlled)